requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
Pillow==10.0.1
Werkzeug==2.3.7
python-magic==0.4.27
requests==2.31.0 